islet_login.sh: test for config after timeout
If TMOUT is reached because a user has not selected
an item in the menu then script the exec's islet_login
and causes a source failure shown below. If that happens
we should just exit the program to be safe.

Choose a number to load the configuration:

1) argus.conf            4) jenkins.conf      7) tcpdump.conf
2) brolive.conf          5) netsniff-ng.conf
3) ids.conf      6) sniffer.conf
[Choice]:
/opt/islet/bin/islet_login: line 26: source: /etc/islet/: is a directory
Are you a new or existing user? [new/existing]:
...
Connection to 127.0.0.1 closed.
please enter the commit message for your changes. Lines starting