add support for service training via port forwarding
We can run the same service on the same port in multiple
containers for various users by creating iptables PREROUTING
rules to create a port forward rules based on the users PPID
and their IP address. Expects uniqueness in the client IP so
it wouldn't work if you're behind a NAT device.

To use set the INTERFACE, BIND, and PORT variables and then
use MOTD to print a message informing the user of how to
connect to the server e.g. MOTD="Connect to http://$SERVER:$PPID"
The PORT variable is to be set to the value of TCP port the
service is listening on in the container. Defaults of
INTERFACE=eth0
BIND="172.17.42.1"
will probably work in most situations.

Also, re-attachment is too much trouble as IP's could change
for users so automatically remove container if the PORT variable
is set.

A cron job runs every minute to remove old PREROUTING rules.