test and create sqlite database
Test for SQLite database and create if it doesn't exist.
If one runs an older version of ISLET which uses a flat
file for credentials it will be removed and a SQLite db
will be created automatically to continue working use
after a ``make update''.

If database doesn't exist then remove user script will
create an empty file with root ownership which cause the
ISLET login script to fail. A fix is check whether the
file is a databse or not first and if it is then remove
the old users