Add support for plugins
Using the same config format with a PLUGIN variable we can
support the executed of plugins from the ISLET menu. Two components
are necessary, the config file that in /etc/islet that lists the plugin
name via PLUGIN which should be set to the filename of the PLUGIN, and
second the plugin should be added to the plugins directory with the
same filename as listed in the config file.

In this commit I added plugins for useful features:
restart - Plugin to send the user back to the islet_shell. From here
they can log in again.
del_user - Plugin to all the user to delete their user account